research_agent/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
newspaper3k
openai
pydantic
python-dotenv
